+++ packages/crdt_socket_sync/CHANGELOG.md
@@ -9,6 +9,10 @@
 - Feature: add `messageCodec` parameter to `WebSocketServer` and `WebSocketClient`
 - Feature: `JsonMessageCodec` now supports `toEncodable` and `reviver` parameters
 - Feature: added out of sync error handling
+- Feature: added `messageBroadcasted` and `messageSent` server events
+
+### Changes
+- Document status request can be sent without a version vector
 
 ## [0.2.0] - 2025-06-26